requirements.txt
flake8
black
jupyter
ipykernel
pandas
numpy
hydra-core
matplotlib
seaborn
tqdm
pytorch-lightning
torchmetrics
transformers==4.26.1
scipy
pytest
fire
jiwer
accelerate